Say Sign when there is nothing to rebase
- The button read Rebase whether it was moving the branch onto a
  base that had moved or only re-signing commits where it already
  sat; it now reads Sign, and Signing while it runs, when signing
  is all there is to do.
- It also read `Rebase ↓↓2`: the count already carries the arrow
  the sidebar uses, and yesterday's change added a second one.
- `PullRequestScope` and its picker move to their own file, which
  the extra lines pushed over the length limit.
